// queryeer-desktop/src/plugins/core.editor/texteditor/types.ts
export type TextRange = {
  startLineNumber: number;
  startColumn: number;
  endLineNumber: number;
  endColumn: number;
};

export type TextLine = {
  readonly lineNumber: number;
  readonly text: string;
  readonly range: TextRange;
};
// ...
export type TextDocument = {
  readonly uri: string;
  readonly languageId: string;
  getText(range?: TextRange): string;
  lineCount: number;
  lineAt(lineNumber: number): TextLine;
};